adding a requests_futures based gbdx_connection
In it's previous form, a variety of metadata requests are executed sequentially through one requests session.  When creating one instance this is a minor inconvenience.  However when instantiating thousands it means that one cannot even multithread to speed things up.  This commit sets up and configures a module-level requests_futures session to enable requests to happen in parallel.