adding support for futures based requests as well as an optional catalog metadata cache
The cache is due to the fact that individual lookups to vectorservices for start to incur a serious performance penalty.  Unlike in the idahoapi case, simply parallelizing the requests doesn't help as it seems that the performance hit is happening server side.  Thus a pre-seeded dict-like cache can be provided.